Add files via upload
1. You like using QR code for your firm/domain, you have QR code on your product. Now you would like to generate code for your products in python using QR code. 
(a). Write a function "EncodeQR" which takes the name of the product(in a string) at most 
5 characters with a minimum of one letter and returns back a list containing the encoded QR code. Use lambda functions wherever you find it appropriate. 
Example if the input is Hi 
H - Ascii value is 72, Binary conversion of 72 is 01001000 as this is the first character pad it with 10000000 so the encoded code for H is1000000001001000 
i-Ascii value is 105,Binary conversion of 105 is 01101001 and this is second character pad it with 01000000 so the encoded code for 'i' is 0100000001101001 
Hi - 10000000010010000100000001101001. 
(b) Put the above function in a module and use another script to import the module and call the function in the script. 
P4:Demonstrate use of object-oriented programming concepts 
You are crazy about your selected domain. Today is The Day you can write a python program on YOUR DOMAIN!!! 
Create a class called "Your Domain Name" which has five attributes(2 string, 3 number). Initialize the variables in the "__init_” method and write at least three more methods. With respect to your domain, you can create methods for getting Input, update the values, verify the condition, and etc,. 
1. Create two objects and call the functions. 
2. Implement the concept of Inheritance and Polymorphism.